requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
pydantic==1.10.9
requests==2.31.0
google-auth==2.20.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
Jinja2==3.1.2
python-multipart==0.0.6
uuid